Generalise method to start app install and move to utils
