+++ app/src/org/commcare/engine/resource/installers/SingleAppInstallation.java
@@ -3,10 +3,11 @@ package org.commcare.engine.resource.installers;
 import org.commcare.CommCareApp;
 import org.commcare.activities.CommCareSetupActivity;
 import org.commcare.engine.resource.AppInstallStatus;
+import org.commcare.engine.resource.ResourceInstallUtils;
 import org.commcare.tasks.ResourceEngineTask;
 
-import static org.commcare.activities.CommCareSetupActivity.handleAppInstallResult;
 import static org.commcare.engine.resource.ResourceInstallUtils.getNewCommCareApp;
+import static org.commcare.engine.resource.ResourceInstallUtils.handleAppInstallResult;
 
 /**
  * Install CC app from the APK's asset directory
@@ -23,29 +24,6 @@ public class SingleAppInstallation {
      * without prompting the user.
      */
     public static void installSingleApp(CommCareSetupActivity activity, int dialogId) {
-        CommCareApp app = getNewCommCareApp();
-
-        ResourceEngineTask<CommCareSetupActivity> task =
-                new ResourceEngineTask<CommCareSetupActivity>(app, dialogId, false, false) {
-                    @Override
-                    protected void deliverResult(CommCareSetupActivity receiver,
-                                                 AppInstallStatus result) {
-                        handleAppInstallResult(this,receiver,result);
-                    }
-
-                    @Override
-                    protected void deliverUpdate(CommCareSetupActivity receiver,
-                                                 int[]... update) {
-                        receiver.updateResourceProgress(update[0][0], update[0][1], update[0][2]);
-                    }
-
-                    @Override
-                    protected void deliverError(CommCareSetupActivity receiver,
-                                                Exception e) {
-                        receiver.failUnknown(AppInstallStatus.UnknownFailure);
-                    }
-                };
-        task.connect(activity);
-        task.executeParallel(SINGLE_APP_REFERENCE);
+        ResourceInstallUtils.startAppInstallAsync(false, dialogId, activity, SINGLE_APP_REFERENCE);
     }
 }
